requirements.txt
streamlit
plotly
pandas
numpy
statsmodels
scikit-learn
scipy
python-dotenv
google-genai
